+++ docs/changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Effort-based Versioning](https://jacobtomlinson.dev
 - Replace `pybtex` by simpler built-in BibTeX parser in `bibsane` written in Lark.
   This eliminates a external dependencies, in particular `latexcodec`, which caused some issues.
 - Only report unused citations in `bibsane`, instead of dropping them.
+- Downgrade `cattrs` dependency to facilitate installation.
 
 ### Fixed